+++ frontend/src/routes/_authenticated/annotation-campaign/$campaignID/phase.$phaseType.module.scss
